add event-listener on abstract
